+++ src/App.js
@@ -1,6 +1,4 @@
-import logo from './logo.svg';
 import './App.css';
-import Login from './login';
 import { Route, Routes, BrowserRouter } from "react-router-dom";
 import Home from './pages/home'
 import Navbar from './navbar';
@@ -8,12 +6,11 @@ import ContactUs from './pages/contactus';
 import SignUp from './pages/signup';
 import Sidebar from './sidebar';
 import {useState} from 'react';
-import AboutUsSection from './AboutUsSection/index';
+import AboutUsSection from './pages/AboutUsSection/index';
+import LoginPage from './pages/login';
 
 function App() {
-  const handleButtonClicked = () => {
-    alert('Hello from React!')
-  }
+  
 
   const [isOpen, setIsOpen] = useState(false)
     
@@ -21,34 +18,8 @@ function App() {
         setIsOpen(!isOpen)
     }
 
-  const LoginPage = () => {
-
-
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <button onClick={handleButtonClicked}>Greetings!</button>
-          <a
-            className="App-link"
-            href="https://github.com/veiserath/Daftcode-frontend"
-            target="_blank"
-            rel="noopener noreferrer">
-            Link to GitHub repo
-          </a>
-          <Login onChange={handleButtonClicked}></Login>
-        </header>
-      </div>
-    );
-  };
-
-
-
-
   return (
     <BrowserRouter>
-      {/* <Route index path='/' element={<LoginPage />} />
-        <Route path='/about' element={<AboutPage />} /> */}
       <Sidebar isOpen={isOpen} toggle={toggle} />
       <Navbar toggle={toggle}/>
         <Routes>
